Revereted changes in build.gradle and package structure
