+++ src/main/java/com/github/csankhala/Application.java
@@ -1,0 +1,18 @@
+package com.github.csankhala.Main;
+
+import io.micronaut.runtime.Micronaut;
+import io.swagger.v3.oas.annotations.*;
+import io.swagger.v3.oas.annotations.info.*;
+
+@OpenAPIDefinition(
+    info = @Info(
+            title = "CustomerService",
+            version = "0.0"
+    )
+)
+public class Application {
+
+    public static void main(String[] args) {
+        Micronaut.run(Application.class, args);
+    }
+}
