--- src/main/java/com/github/csankhala/Main/Entity/Customer.java
@@ -1,42 +1,0 @@
-package com.github.csankhala.Main.Entity;
-
-import java.util.Objects;
-
-public class Customer {
-
-    private final int id;
-    private final String name;
-    private final String address;
-
-
-    public Customer(int id, String name, String address) {
-        this.id = id;
-        this.name = name;
-        this.address = address;
-    }
-
-    public String getAddress() {
-        return address;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public int getId() {
-        return id;
-    }
-
-    @Override
-    public boolean equals(Object o) {
-        if (this == o) return true;
-        if (!(o instanceof Customer)) return false;
-        Customer customer = (Customer) o;
-        return getId() == customer.getId();
-    }
-
-    @Override
-    public int hashCode() {
-        return Objects.hash(getId());
-    }
-}
